moving bookmarks and centeralizing json loader
