+++ src/modules/cmdpal/src/WindowsCommandPalette/Builtins/Bookmark/OpenInTerminalAction.cs
@@ -1,0 +1,48 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Text.Json;
+using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
+using Microsoft.Windows.CommandPalette.Extensions;
+using Microsoft.Windows.CommandPalette.Extensions.Helpers;
+using Windows.Foundation;
+using Windows.System;
+using static System.Runtime.InteropServices.JavaScript.JSType;
+
+namespace Run.Bookmarks;
+
+internal sealed class OpenInTerminalAction : InvokableCommand
+{
+    private readonly string _folder;
+
+    public OpenInTerminalAction(string folder)
+    {
+        Name = "Open in Terminal";
+        _folder = folder;
+    }
+
+    public override ICommandResult Invoke()
+    {
+        try
+        {
+            // Start Windows Terminal with the specified folder
+            var startInfo = new System.Diagnostics.ProcessStartInfo
+            {
+                FileName = "wt.exe",
+                Arguments = $"-d \"{_folder}\"",
+                UseShellExecute = true,
+            };
+            System.Diagnostics.Process.Start(startInfo);
+        }
+        catch (Exception ex)
+        {
+            System.Diagnostics.Debug.WriteLine($"Error launching Windows Terminal: {ex.Message}");
+        }
+
+        return ActionResult.Dismiss();
+    }
+}
