+++ src/modules/cmdpal/src/WindowsCommandPalette/Builtins/Bookmark/UrlAction.cs
@@ -1,0 +1,105 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+using System.Text.Json;
+using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
+using Microsoft.Windows.CommandPalette.Extensions;
+using Microsoft.Windows.CommandPalette.Extensions.Helpers;
+using Windows.Foundation;
+using Windows.System;
+using static System.Runtime.InteropServices.JavaScript.JSType;
+
+namespace Run.Bookmarks;
+
+public class UrlAction : InvokableCommand
+{
+    private bool IsContainsPlaceholder => _url.Contains('{') && _url.Contains('}');
+
+    public string Type { get; }
+
+    public string Url { get; }
+
+    private readonly string _url;
+
+    public UrlAction(string name, string url, string type)
+    {
+        _url = url;
+        Icon = new(IconFromUrl(_url, type));
+        Name = name;
+        Type = type;
+        Url = url;
+    }
+
+    public override ActionResult Invoke()
+    {
+        var target = _url;
+        try
+        {
+            Uri? uri = GetUri(target);
+            if (uri != null)
+            {
+                _ = Launcher.LaunchUriAsync(uri);
+            }
+            else
+            {
+                // throw new UriFormatException("The provided URL is not valid.");
+            }
+        }
+        catch (Exception ex)
+        {
+            System.Diagnostics.Debug.WriteLine($"Error launching URL: {ex.Message}");
+        }
+
+        return ActionResult.Dismiss();
+    }
+
+    internal static Uri? GetUri(string url)
+    {
+        Uri? uri;
+        if (!Uri.TryCreate(url, UriKind.Absolute, out uri))
+        {
+            if (!Uri.TryCreate("https://" + url, UriKind.Absolute, out uri))
+            {
+                return null;
+            }
+        }
+
+        return uri;
+    }
+
+    internal static string IconFromUrl(string url, string type)
+    {
+        switch (type)
+        {
+            case "file":
+                return "📄";
+            case "folder":
+                return "📁";
+            case "web":
+            default:
+                // Get the base url up to the first placeholder
+                var placeholderIndex = url.IndexOf('{');
+                var baseString = placeholderIndex > 0 ? url.Substring(0, placeholderIndex) : url;
+                try
+                {
+                    Uri? uri = GetUri(baseString);
+                    if (uri != null)
+                    {
+                        var hostname = uri.Host;
+                        var faviconUrl = $"{uri.Scheme}://{hostname}/favicon.ico";
+                        return faviconUrl;
+                    }
+                }
+                catch (UriFormatException)
+                {
+                    // return "🔗";
+                }
+
+                return "🔗";
+        }
+    }
+}
